code for LayernormSimpleRNN moved into cell.py and cell_test.py
