+++ tensorflow_addons/rnn/BUILD
@@ -7,7 +7,6 @@ py_library(
     srcs = [
         "__init__.py",
         "cell.py",
-        "layernorm_simplernn.py",
     ],
     deps = [
         "//tensorflow_addons/utils",
@@ -18,15 +17,6 @@ py_test(
     name = "cell_test",
     size = "small",
     srcs = ["cell_test.py"],
-    deps = [
-        ":rnn",
-    ],
-)
-
-py_test(
-    name = "layernorm_simplernn_test",
-    size = "small",
-    srcs = ["layernorm_simplernn_test.py"],
     deps = [
         ":rnn",
     ],
